requirements_dev.txt
pip
bumpversion
wheel
watchdog
flake8
tox
coverage
Sphinx
sphinx-click
twine

pytest
pytest-runner
